+++ src/Commands/BarionCommand.php
@@ -1,12 +1,14 @@
 <?php
 
-namespace VendorName\Skeleton\Commands;
+declare(strict_types=1);
+
+namespace LovaszCC\Barion\Commands;
 
 use Illuminate\Console\Command;
 
-class SkeletonCommand extends Command
+final class BarionCommand extends Command
 {
-    public $signature = 'skeleton';
+    public $signature = 'laravel-barion';
 
     public $description = 'My command';