+++ src/Enums/Locale.php
@@ -1,0 +1,24 @@
+<?php
+
+declare(strict_types=1);
+
+namespace LovaszCC\Barion\Enums;
+
+enum Locale: string
+{
+    public const CZ = 'cs-CZ';
+
+    public const DE = 'de-DE';
+
+    public const EN = 'en-US';
+
+    public const ES = 'es-ES';
+
+    public const FR = 'fr-FR';
+
+    public const HU = 'hu-HU';
+
+    public const SK = 'sk-SK';
+
+    public const SI = 'sl-SI';
+}