+++ CHANGELOG.md
@@ -1,3 +1,11 @@
 # Changelog
 
-All notable changes to `:package_name` will be documented in this file.
+v1.0.0 - Initial release
+
+-   Initial release
+-   Added Barion class
+-   Added Barion facade
+-   Added Barion config
+-   Added Barion enums
+-   Added Barion tests
+-   Added Barion documentation